+++ app/components/settings/categories/general-setting/AboutYoroiSettingsBlock.scss
