+++ app/components/settings/categories/general-setting/DisplaySettingsBlock.scss
