+++ app/components/settings/categories/general-setting/GeneralSettings.scss
@@ -1,4 +1,4 @@
-@import '../../../themes/mixins/error-message';
+@import '../../../../themes/mixins/error-message';
 
 .component {
   margin-bottom: 20px;
